Revert "+ frequency is float value now"
This reverts commit 4bde59333fa5d82c2e863e9c334581fc08af7ce3.

+++ YamahaTuner.py
@@ -50,7 +50,7 @@ class YamahaTuner:
     def frequency(self):
         return self._frequencies[self._band]
 
-    def set_frequency(self, freq: float):
+    def set_frequency(self, freq: int):
         self._frequencies[self._band] = freq
 
     def next_preset(self):
